+++ package.json
@@ -69,8 +69,8 @@
     "@n8tb1t/use-scroll-position": "^2.0.3",
     "@reduxjs/toolkit": "1.8.1",
     "@segment/analytics-next": "^1.43.0",
-    "antd": "^4.23.6",
     "ansi-to-react": "^6.1.6",
+    "antd": "^4.23.6",
     "axios": "0.27.2",
     "classnames": "2.3.1",
     "date-fns": "^2.28.0",
@@ -92,6 +92,7 @@
     "react-transition-group": "4.4.2",
     "react-use-websocket": "^4.2.0",
     "redux-logger": "3.0.6",
+    "strip-ansi": "^7.0.1",
     "styled-components": "5.3.5",
     "uuid": "8.3.2",
     "web-vitals": "2.1.4"
