+++ src/antd-theme/my-antd-theme.less
@@ -55,5 +55,3 @@
 @notification-bg: #1e293b;
 // dropdown
 @dropdown-menu-bg: #1e293b;
-// popover
-@popover-background: #0e6ff7;
