+++ src/components/atoms/CopyButton/index.ts
@@ -1,0 +1,1 @@
+export {default} from './CopyButton';
