+++ src/components/atoms/TestkubeTheme/AnsiClassesMapping.ts
@@ -1,0 +1,82 @@
+import Colors from '@styles/Colors';
+
+export default `
+  .ansi-black-fg {
+    color: ${Colors.slate600};
+  }
+  .ansi-bright-black-fg {
+    color: ${Colors.slate400};
+  }
+  .ansi-black-bg {
+    color: ${Colors.slate200};
+  }
+
+  .ansi-white-fg {
+    color: ${Colors.slate400};
+  }
+  .ansi-bright-white-fg {
+    color: ${Colors.slate200};
+  }
+  .ansi-white-bg {
+    color: ${Colors.slate200};
+  }
+
+  .ansi-red-fg {
+    color: ${Colors.rose500};
+  }
+  .ansi-bright-red-fg {
+    color: ${Colors.rose300};
+  }
+  .ansi-red-bg {
+    color: ${Colors.rose900};
+  }
+
+  .ansi-green-fg {
+    color: ${Colors.lime500};
+  }
+  .ansi-bright-green-fg {
+    color: ${Colors.lime300};
+  }
+  .ansi-green-bg {
+    color: ${Colors.lime900};
+  }
+
+  .ansi-yellow-fg {
+    color: ${Colors.yellow500};
+  }
+  .ansi-bright-yellow-fg {
+    color: ${Colors.yellow300};
+  }
+  .ansi-yellow-bg {
+    color: ${Colors.yellow900};
+  }
+
+  .ansi-blue-fg {
+    color: ${Colors.sky500};
+  }
+  .ansi-bright-blue-fg {
+    color: ${Colors.sky300};
+  }
+  .ansi-blue-bg {
+    color: ${Colors.sky900};
+  }
+
+  .ansi-magenta-fg {
+    color: ${Colors.pink500};
+  }
+  .ansi-bright-magenta-fg {
+    color: ${Colors.pink300};
+  }
+  .ansi-magenta-bg {
+    color: ${Colors.pink900};
+  }
+
+  .ansi-cyan-fg {
+    color: ${Colors.cyan500};
+  }
+  .ansi-bright-cyan-fg {
+    color: ${Colors.cyan300};
+  }
+  .ansi-cyan-bg {
+    color: ${Colors.cyan900};
+  }`;
