+++ src/components/atoms/index.ts
@@ -12,3 +12,6 @@ export {default as ProtectedRoute} from './ProtectedRoute';
 export {default as Pre} from './Pre';
 export {default as CreatableMultiSelect} from './CreatableMultiSelect';
 export {default as SplittedLabelText} from './SplittedLabelText';
+export {default as CopyButton} from './CopyButton';
+export {default as AnsiClassesMapping} from './TestkubeTheme/AnsiClassesMapping';
+export {default as TestkubeCodeTheme} from './TestkubeTheme/CodeTheme';
