+++ src/components/molecules/ExecutionDetails/ExecutionDetails.styled.ts
@@ -84,10 +84,7 @@ export const StyledAntTabs = styled(Tabs)``;
 
 export const StyledAntTabPane = styled(Tabs.TabPane)``;
 
-export const StyledTestExecutionDetailsTabsContainer = styled.div`
-  height: 100%;
-  overflow: auto;
-`;
+export const StyledTestExecutionDetailsTabsContainer = styled.div``;
 
 export const StyledTestSuiteExecutionDetailsTabsContainer = styled.div``;
