+++ src/components/molecules/LogOutput/LogOutput.styled.tsx
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 
+import {AnsiClassesMapping} from '@atoms';
+
 import Colors from '@styles/Colors';
 import {invisibleScroll} from '@styles/globalStyles';
 
@@ -14,10 +16,11 @@ export const StyledLogOutputContainer = styled.div<{$isFullScreen?: boolean}>`
     `
       : ''};
 
-  // overflow-x: auto;
-
   display: flex;
   flex-direction: column;
+
+  overflow: auto;
+  height: 70vh;
 
   border-radius: ${({$isFullScreen}) => ($isFullScreen ? '0' : '4px')};
 
@@ -71,86 +74,7 @@ export const StyledPreLogText = styled.pre`
   font-size: 12px;
 
   ${invisibleScroll}
-
-  .ansi-black-fg {
-    color: ${Colors.slate600};
-  }
-  .ansi-bright-black-fg {
-    color: ${Colors.slate400};
-  }
-  .ansi-black-bg {
-    color: ${Colors.slate200};
-  }
-
-  .ansi-white-fg {
-    color: ${Colors.slate400};
-  }
-  .ansi-bright-white-fg {
-    color: ${Colors.slate200};
-  }
-  .ansi-white-bg {
-    color: ${Colors.slate200};
-  }
-
-  .ansi-red-fg {
-    color: ${Colors.rose500};
-  }
-  .ansi-bright-red-fg {
-    color: ${Colors.rose300};
-  }
-  .ansi-red-bg {
-    color: ${Colors.rose900};
-  }
-
-  .ansi-green-fg {
-    color: ${Colors.lime500};
-  }
-  .ansi-bright-green-fg {
-    color: ${Colors.lime300};
-  }
-  .ansi-green-bg {
-    color: ${Colors.lime900};
-  }
-
-  .ansi-yellow-fg {
-    color: ${Colors.yellow500};
-  }
-  .ansi-bright-yellow-fg {
-    color: ${Colors.yellow300};
-  }
-  .ansi-yellow-bg {
-    color: ${Colors.yellow900};
-  }
-
-  .ansi-blue-fg {
-    color: ${Colors.sky500};
-  }
-  .ansi-bright-blue-fg {
-    color: ${Colors.sky300};
-  }
-  .ansi-blue-bg {
-    color: ${Colors.sky900};
-  }
-
-  .ansi-magenta-fg {
-    color: ${Colors.pink500};
-  }
-  .ansi-bright-magenta-fg {
-    color: ${Colors.pink300};
-  }
-  .ansi-magenta-bg {
-    color: ${Colors.pink900};
-  }
-
-  .ansi-cyan-fg {
-    color: ${Colors.cyan500};
-  }
-  .ansi-bright-cyan-fg {
-    color: ${Colors.cyan300};
-  }
-  .ansi-cyan-bg {
-    color: ${Colors.cyan900};
-  }
+  ${AnsiClassesMapping}
 `;
 
 export const StyledLogOutputActionsContainer = styled.ul`
@@ -168,7 +92,7 @@ export const StyledLogOutputHeaderContainer = styled.div`
   display: flex;
   justify-content: space-between;
 
-  padding: 8px;
+  padding: 12px;
   border-radius: inherit;
 
   background: ${Colors.slate900};
